backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
psycopg2-binary
gunicorn
requests
google-generativeai
